merge overlapping intervals
In-Place Merging – O(n*log(n)) time and O(1) space
The above solution requires O(n) space for sorting the merged intervals. We can avoid the extra space by doing merge operations in place.  

Follow the steps mentioned below to implement the approach:

Sort all intervals in increasing order of start time.
Traverse sorted intervals starting from the first interval, 
Do the following for every interval.
If the current interval is not the first interval and it overlaps with the previous interval, then merge it with the previous interval. Keep doing it while the interval overlaps with the previous one.         
If the current interval does not overlap with the previous interval, move to the next interval.